requirements.txt
Django==1.9.2
django-debug-toolbar==1.4
mysqlclient==1.3.7
django-bower==5.1.0
djangorestframework==3.3.3
djangorestframework-csv==1.4.1
django-widget-tweaks==1.4.1
factory-boy==2.6.1
